fix: remove old footer and duplicate navbar imports
