+++ src/components/Layout/Footer.js
@@ -1,371 +1,155 @@
-import { useState } from "react";
+import React from "react";
 import {
-  FaInfoCircle,
-  FaInstagram,
-  FaDiscord,
-  FaTelegram,
   FaGithub,
-  FaQuestionCircle,
-  FaEnvelope,
-  FaBookOpen,
-  FaPlus,
-  // FaClipboardList,  //can be reactivated later
-  FaUsers,
-  FaBook,
-  // FaServer,   //can be reactivated later
-  FaHome,
-  FaCalendarAlt,
-  FaStar,
-  FaFolder,
-  FaTrophy,
-  FaComments,
   FaLinkedin,
+  FaDiscord,
+  FaTelegramPlane,
+  FaInstagram,
+  FaEnvelope,
 } from "react-icons/fa";
-import { Link } from "react-router-dom";
 
-
-// Toast Component
-const Toast = ({ message, type, onClose }) => {
-  return (
-    <div
-      className={`fixed bottom-4 left-4 p-4 rounded-md shadow-lg text-white ${
-        type === "success" ? "bg-green-500" : "bg-red-500"
-      } transition-opacity duration-300`}
-    >
-      <div className="flex items-center justify-between">
-        <span>{message}</span>
-        <button
-          onClick={onClose}
-          className="ml-4 text-white hover:text-gray-200"
-        >
-          &times;
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const Footer = () => {
+const NewFooter = () => {
   const currentYear = new Date().getFullYear();
-  const [email, setEmail] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [toast, setToast] = useState({ show: false, message: "", type: "" });
-
-  const showToast = (message, type = "success") => {
-    setToast({ show: true, message, type });
-    setTimeout(() => {
-      setToast({ show: false, message: "", type: "" });
-    }, 3000);
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!email) {
-      showToast("Please enter your email address", "error");
-      return;
-    }
-
-    if (!/\S+@\S+\.\S+/.test(email)) {
-      showToast("Please enter a valid email address", "error");
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      // Simulate API call
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      // In a real application, you would make an API call to your backend here
-      console.log("Submitting email:", email);
-
-      showToast("Thank you for subscribing to our newsletter!");
-      setEmail("");
-    } catch (error) {
-      showToast("Something went wrong. Please try again.", "error");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const footerLinks = {
-    quick_links: [
-      { name: "Home", href: "/", icon: <FaHome size={14} /> },
-      { name: "Events", href: "/events", icon: <FaCalendarAlt size={14} /> },
-      { name: "Hackathons", href: "/hackathons", icon: <FaStar size={14} /> },
-      { name: "Projects", href: "/projects", icon: <FaFolder size={14} /> },
-      { name: "About", href: "/about", icon: <FaInfoCircle size={14} /> },
-    ],
-    community: [
-      {
-        name: "Create Event",
-        href: "/create-event",
-        icon: <FaPlus size={14} />,
-      },
-      {
-        name: "Community Events",
-        href: "/communityEvent",
-        icon: <FaUsers size={14} />,
-      },
-      {
-        name: "Documentation",
-        href: "/documentation",
-        icon: <FaBook size={14} />,
-      },
-      {
-        name: "Contributors",
-        href: "/contributors",
-        icon: <FaUsers size={14} />,
-      },
-      {
-        name: "Contributors Guide",
-        href: "/contributorguide",
-        icon: <FaBook size={14} />,
-      },
-      {
-        name: "LeaderBoard",
-        href: "/leaderBoard",
-        icon: <FaTrophy size={14} />,
-      },
-    ],
-    support: [
-      {
-        name: "Help Center",
-        href: "/helpcenter",
-        icon: <FaQuestionCircle size={14} />,
-      },
-      { name: "Contact Us", href: "/contact", icon: <FaEnvelope size={14} /> },
-      { name: "Feedback", href: "/feedback", icon: <FaComments size={14} /> },
-      { name: "API Docs", href: "/apiDocs", icon: <FaBookOpen size={14} /> },
-    ],
-  };
-
-  const legalLinks = [
-    { name: "Privacy Policy", href: "/privacy" },
-    { name: "Terms of Service", href: "/terms" },
-  ];
-
-  const socialLinks = [
-    {
-      name: "GitHub",
-      href: "https://github.com/sandeepvashishtha/Eventra",
-      icon: (
-        <FaGithub
-          className="size-10 p-2 rounded-full text-black bg-white shadow-sm hover:shadow-lg transition-all duration-300 hover:bg-gray-100 hover:scale-110 hover:-translate-y-1"
-          size={20}
-        />
-      ),
-    },
-
-    {
-      name: "LinkedIn",
-      href: "https://www.linkedin.com/in/sandeepvashishtha/",
-      icon: (
-        <FaLinkedin
-          className="size-10 p-2 rounded-full text-black bg-white shadow-sm hover:shadow-lg transition-all duration-300 hover:bg-gray-100 hover:scale-110 hover:-translate-y-1"
-          size={20}
-        />
-      ),
-    },
-
-    {
-      name: "Discord",
-      href: "https://www.discord.com/",
-      icon: (
-        <FaDiscord
-          className="size-10 p-2 rounded-full text-black bg-white shadow-sm hover:shadow-lg transition-all duration-300 hover:bg-gray-100 hover:scale-110 hover:-translate-y-1"
-          size={20}
-        />
-      ),
-    },
-
-    {
-      name: "Telegram",
-      href: "https://www.telegram.com/",
-      icon: (
-        <FaTelegram
-          className="size-10 p-2 rounded-full text-black bg-white shadow-sm hover:shadow-lg transition-all duration-300 hover:bg-gray-100 hover:scale-110 hover:-translate-y-1"
-          size={20}
-        />
-      ),
-    },
-
-    {
-      name: "Instagram",
-      href: "https://www.instagram.com/",
-      icon: (
-        <FaInstagram
-          className="size-10 p-2 rounded-full text-black bg-white shadow-sm hover:shadow-lg transition-all duration-300 hover:bg-gray-100 hover:scale-110 hover:-translate-y-1"
-          size={20}
-        />
-      ),
-    },
-
-    
-  ];
 
   return (
-    <>
-      {/* UPDATED: Added dark mode background and border colors */}
-      <footer 
-        className="bg-white dark:bg-gray-900 border-t border-gray-100 dark:border-gray-800"
-        // AOS Implementation
-        data-aos="fade-up"
-        data-aos-duration="1000"
-        data-aos-offset="100" // Start animation a bit earlier
-        // End AOS Implementation
-      >
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 pt-12 pb-6">
-          <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-5 gap-6">
-            <div 
-              className="space-y-4 lg:col-span-2"
-              // AOS Implementation - Logo/Newsletter Column
-              data-aos="fade-up"
-              data-aos-delay="0"
-            >
-              {/* UPDATED: Added dark mode text color */}
-              <h2
-                className="text-2xl sm:text-3xl font-bold text-black dark:text-white"
-                style={{ fontFamily: "Anton, sans-serif" }}
-              >
+    <footer className="bg-white border-t border-gray-100 mt-20">
+      <div className="max-w-7xl mx-auto px-6 lg:px-10 py-12">
+        
+        {/* MAIN GRID */}
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8 lg:gap-12">
+          
+          {/* LEFT SIDE: Brand, Description & Socials */}
+          <div>
+            <div className="flex items-center gap-2 mb-4">
+              <span className="text-3xl font-bold text-black tracking-tight">
                 Eventra
-              </h2>
-              {/* UPDATED: Added dark mode text color */}
-              <p className="text-gray-600 dark:text-gray-400 text-sm">
-                Open-source event management for communities worldwide.
-              </p>
-
-              {/* Newsletter Subscription Form */}
-              <div className="mt-4">
-                {/* UPDATED: Added dark mode text color */}
-                <h4 className="text-sm font-semibold text-gray-900 dark:text-gray-100 uppercase tracking-wider mb-2">
-                  Subscribe to our newsletter
-                </h4>
-                {/* UPDATED: Added dark mode text color */}
-                <p className="text-gray-600 dark:text-gray-400 text-sm mb-2">
-                  Get the latest updates, event tips, and community news.
-                </p>
-                <form
-                  onSubmit={handleSubmit}
-                  className="flex flex-col sm:flex-row gap-3"
-                >
-                  <div className="relative flex-grow">
-                    <FaEnvelope className="absolute left-3 top-1/2 -translate-y-1/2 text-gray-400" />
-                    <input
-                      type="email"
-                      value={email}
-                      onChange={(e) => setEmail(e.target.value)}
-                      placeholder="Enter your email"
-                      className="pl-10 pr-4 py-2.5 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 border border-gray-300 dark:border-gray-600 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent w-full"
-                      disabled={isSubmitting}
-                    />
-                  </div>
-                  <button
-                    type="submit"
-                    disabled={isSubmitting}
-                    className="w-full sm:w-auto px-4 py-2.5 bg-black text-white dark:bg-white dark:text-black dark:hover:bg-gray-400 rounded-md hover:bg-gray-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                  >
-                    {isSubmitting ? "Subscribing..." : "Subscribe"}
-                  </button>
-                </form>
-                {/* UPDATED: Added dark mode text color */}
-                <p className="text-xs text-gray-500 dark:text-gray-400 mt-1">
-                  We respect your privacy. Unsubscribe at any time.
-                </p>
-              </div>
-              {/* Social Media Icons - Below Newsletter */}
-              <div className="mt-6">
-                <h4 className="text-sm font-semibold text-gray-900 dark:text-gray-100 uppercase tracking-wider mb-3">
-                  Follow Us
-                </h4>
-                <div className="flex flex-wrap gap-3">
-                  {socialLinks.map((link) => (
-                    <a
-                      key={link.name}
-                      href={link.href}
-                      className="text-gray-400 dark:text-gray-500 hover:text-indigo-600 dark:hover:text-indigo-400 transition-colors"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      aria-label={link.name}
-                      title={link.name}
-                    >
-                      <span className="sr-only">{link.name}</span>
-                      {link.icon}
-                    </a>
-                  ))}
-                </div>
-              </div>
+              </span>
             </div>
 
-            {Object.entries(footerLinks).map(([key, links]) => (
-              <div 
-                key={key} 
-                className="py-2"
-                data-aos="fade-up"
-                data-aos-delay={key === "quick_links" ? "100" : key === "community" ? "200" : "300"}
-              >
-                <h4 className="text-xs font-bold text-gray-900 dark:text-gray-100 uppercase tracking-widest mb-6">
-                  {key.replace("_", " ")}
-                </h4>
-                <ul className="space-y-4">
-                  {links.map((link) => (
-                    <li key={link.name}>
-                      <Link
-                        to={link.href}
-                        className="text-sm text-gray-600 hover:text-black flex items-center gap-4 transition-all duration-300 hover:translate-x-1 group"
-                      >
-                        {link.icon && (
-                          <span className="text-black group-hover:text-black group-hover:scale-110 transition-all duration-300">
-                            {link.icon}
-                          </span>
-                        )}
-                        <span>
-                          {link.name}
-                        </span>
-                      </Link>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            ))}
+            <p className="text-gray-500 leading-relaxed text-sm mb-4 max-w-sm">
+              Open-source event management for communities worldwide.
+            </p>
+
+            {/* SOCIAL ICONS */}
+            <div className="flex gap-4 text-black text-lg">
+              <a href="#" className="hover:text-gray-500 transition" aria-label="GitHub">
+                <FaGithub />
+              </a>
+              <a href="#" className="hover:text-gray-500 transition" aria-label="LinkedIn">
+                <FaLinkedin />
+              </a>
+              <a href="#" className="hover:text-gray-500 transition" aria-label="Discord">
+                <FaDiscord />
+              </a>
+              <a href="#" className="hover:text-gray-500 transition" aria-label="Telegram">
+                <FaTelegramPlane />
+              </a>
+              <a href="#" className="hover:text-gray-500 transition" aria-label="Instagram">
+                <FaInstagram />
+              </a>
+            </div>
           </div>
 
-          {/* Copyright - Below the line, centered */}
-          {/* UPDATED: Added dark mode border color */}
-          <div className="border-t border-gray-200 dark:border-gray-800 mt-12 pt-8 flex flex-col md:flex-row justify-between items-center gap-6">
-            <p className="text-sm text-gray-500 dark:text-gray-400 text-center md:text-left">
-              © {currentYear} Eventra. All rights reserved. Created with ❤️ by Sandeep Vashishtha, Rhythm and the amazing open-source community.
+          {/* QUICK LINKS */}
+          <div>
+            <h3 className="text-[15px] font-semibold text-black mb-4 uppercase tracking-wider">
+              Quick Links
+            </h3>
+            <ul className="space-y-3 text-sm text-gray-500">
+              <li className="hover:text-black cursor-pointer transition">Home</li>
+              <li className="hover:text-black cursor-pointer transition">Events</li>
+              <li className="hover:text-black cursor-pointer transition">Hackathons</li>
+              <li className="hover:text-black cursor-pointer transition">Projects</li>
+              <li className="hover:text-black cursor-pointer transition">About</li>
+            </ul>
+          </div>
+
+          {/* SUPPORT LINKS */}
+          <div>
+            <h3 className="text-[15px] font-semibold text-black mb-4 uppercase tracking-wider">
+              Support
+            </h3>
+            <ul className="space-y-3 text-sm text-gray-500">
+              <li className="hover:text-black cursor-pointer transition">Help Center</li>
+              <li className="hover:text-black cursor-pointer transition">Contact Us</li>
+              <li className="hover:text-black cursor-pointer transition">Feedback</li>
+              <li className="hover:text-black cursor-pointer transition">API Docs</li>
+            </ul>
+          </div>
+
+          {/* COMMUNITY LINKS */}
+          <div>
+            <h3 className="text-[15px] font-semibold text-black mb-4 uppercase tracking-wider">
+              Community
+            </h3>
+            <ul className="space-y-3 text-sm text-gray-500">
+              <li className="hover:text-black cursor-pointer transition">Create Event</li>
+              <li className="hover:text-black cursor-pointer transition">Community Events</li>
+              <li className="hover:text-black cursor-pointer transition">Documentation</li>
+              <li className="hover:text-black cursor-pointer transition">Contributors</li>
+              <li className="hover:text-black cursor-pointer transition">Contributors Guide</li>
+              <li className="hover:text-black cursor-pointer transition">LeaderBoard</li>
+            </ul>
+          </div>
+
+        </div>
+
+        {/* MIDDLE SECTION: TIGHTLY FITTED NEWSLETTER BOX */}
+        <div className="mt-6 flex justify-center">
+          <div className="w-full max-w-lg bg-gray-50 border border-gray-100 rounded-xl p-4 sm:p-5 flex flex-col items-center text-center">
+            <h4 className="text-xs font-bold uppercase tracking-wider text-black mb-1">
+              Subscribe to our newsletter
+            </h4>
+            <p className="text-xs text-gray-500 mb-4">
+              Get the latest updates, event tips, and community news.
             </p>
-            <div className="flex gap-6">
-              {legalLinks.map((link) => (
-                <Link
-                  key={link.name}
-                  to={link.href}
-                  className="text-xs text-gray-500 hover:text-indigo-600 dark:hover:text-indigo-400 transition-colors"
-                >
-                  {link.name}
-                </Link>
-              ))}
-            </div>
+            
+            <form onSubmit={(e) => e.preventDefault()} className="flex flex-col sm:flex-row gap-3 mb-2 w-full justify-center">
+              <div className="relative flex-1 max-w-xs">
+                <span className="absolute inset-y-0 left-3 flex items-center text-gray-400">
+                  <FaEnvelope size={14} />
+                </span>
+                <input
+                  type="email"
+                  placeholder="Enter your email"
+                  className="w-full pl-9 pr-3 py-2 bg-white border border-gray-200 rounded-lg text-sm text-gray-800 placeholder-gray-400 outline-none focus:border-gray-400 transition"
+                  required
+                />
+              </div>
+              <button
+                type="submit"
+                className="bg-black text-white px-5 py-2 rounded-lg text-sm font-medium hover:bg-gray-800 transition shadow-sm whitespace-nowrap"
+              >
+                Subscribe
+              </button>
+            </form>
+            
+            <p className="text-[11px] text-gray-400">
+              We respect your privacy. Unsubscribe at any time.
+            </p>
           </div>
         </div>
-      </footer>
 
-      {/* Toast Notification */}
-      {toast.show && (
-        <Toast
-          message={toast.message}
-          type={toast.type}
-          onClose={() => setToast({ show: false, message: "", type: "" })}
-        />
-      )}
-    </>
+        {/* BOTTOM BAR */}
+        <div className="border-t border-gray-100 mt-8 pt-6 flex flex-col sm:flex-row justify-between items-center gap-4 text-xs text-gray-400">
+          <p>© {currentYear} Eventra. All rights reserved. Created by Sandeep Vashishtha</p>
+
+          <div className="flex gap-6">
+            <a href="#" className="hover:text-black underline underline-offset-4 transition">
+              Privacy Policy
+            </a>
+            <a href="#" className="hover:text-black underline underline-offset-4 transition">
+              Terms of Service
+            </a>
+            <a href="#" className="hover:text-black underline underline-offset-4 transition">
+              Cookies Settings
+            </a>
+          </div>
+        </div>
+
+      </div>
+    </footer>
   );
 };
 
-export default Footer;
-
-
-
-
+export default NewFooter;
